refactor(workspace): extract kanban-backend and kanban-api out of kanban-service
Moves the KanbanBackend/RemoteWrites traits into a new kanban-backend crate
and the HTTP wire DTOs into a new kanban-api crate, so backend implementations
depend on a low-level abstraction rather than on the orchestrator.

kanban-service re-exports both, so no consumer source changes are required.
kanban-backend-http drops its kanban-service dependency entirely, which breaks
the cycle that made it impossible for kanban-service to reference a remote
backend.

test_support moves out of the #[cfg(test)] block into a test-helpers gated
module, since a cfg(test) module is not visible across a crate boundary. The
SqliteBackend lifecycle tests move to sqlite_backend.rs, where the type they
exercise still lives; left in place they would have silently dead-stripped.

The factory compile-lock guard's globs are repointed at the new api location
in this same commit -- stale globs match nothing and the guard passes vacuously.
