+++ Cargo.toml
@@ -6,6 +6,8 @@ members = [
     "crates/kanban-persistence",
     "crates/kanban-persistence-json",
     "crates/kanban-persistence-sqlite",
+    "crates/kanban-api",
+    "crates/kanban-backend",
     "crates/kanban-service",
     "crates/kanban-backend-http",
     "crates/kanban-tui",
