+++ crates/kanban-api/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "kanban-api"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Wire-format DTOs for the kanban HTTP API, shared by kanban-server and HTTP backend clients"
+
+[lib]
+name = "kanban_api"
+path = "src/lib.rs"
+
+[features]
+schemars = ["dep:schemars"]
+
+[dependencies]
+kanban-core = { path = "../kanban-core", version = "^0.7" }
+kanban-domain = { path = "../kanban-domain", version = "^0.7" }
+serde.workspace = true
+serde_json.workspace = true
+uuid.workspace = true
+chrono.workspace = true
+schemars = { version = "1.0", optional = true, features = ["uuid1"] }
